expose `ErgoBox::serialized_additional_registers()` in Wasm;
extract additional registers serialization;
